Move serde code into separate deserialize module
